refactor(adapter): cut over start to lifecycle verbs and seal trust boundary (F.5)
